third_party/WebKit/LayoutTests/fast/gradients/unprefixed-gradient-parsing-expected.txt

linear-gradient

PASS testGradient("background-image: linear-gradient(black 0%)") is "none"
PASS testGradient("background-image: linear-gradient(top)") is "none"
PASS testGradient("background-image: linear-gradient(to top)") is "none"
PASS testGradient("background-image: linear-gradient(10deg)") is "none"
PASS testGradient("background-image: linear-gradient(10deg black, white)") is "none"
PASS testGradient("background-image: linear-gradient(to black, white)") is "none"
PASS testGradient("background-image: linear-gradient(to left black, white)") is "none"
PASS testGradient("background-image: linear-gradient(to top center, black 0%, white)") is "none"
PASS testGradient("background-image: linear-gradient(10px 20%, black 0%)") is "none"
PASS testGradient("background-image: linear-gradient(to left left)") is "none"
PASS testGradient("background-image: linear-gradient(to left center)") is "none"
PASS testGradient("background-image: linear-gradient(to left 10deg, black 0%, white)") is "none"
PASS testGradient("background-image: linear-gradient(black)") is "none"
PASS testGradient("background-image: linear-gradient(10deg, black 0%, 25%, 50%, white)") is "none"
PASS testGradient("background-image: linear-gradient(10deg, black 0%, 50%)") is "none"
PASS testGradient("background-image: linear-gradient(10deg, 50%, white)") is "none"
PASS testGradient("background-image: linear-gradient(10deg, 0%, black, white)") is "none"
PASS testGradient("background-image: linear-gradient(10deg, white, black, 80%)") is "none"
PASS testGradient("background-image: linear-gradient(10deg, black, , white)") is "none"
PASS testGradient("background-image: linear-gradient(black, white)") is "linear-gradient(black, white)"
PASS testGradient("background-image: linear-gradient(black 0, white)") is "linear-gradient(black 0px, white)"
PASS testGradient("background-image: linear-gradient(black 0%, white)") is "linear-gradient(black 0%, white)"
PASS testGradient("background-image: linear-gradient(black 0%, white)") is "linear-gradient(black 0%, white)"
PASS testGradient("background-image: linear-gradient(black 10px, white 20px)") is "linear-gradient(black 10px, white 20px)"
PASS testGradient("background-image: linear-gradient(to top left, black 0%, white)") is "linear-gradient(to left top, black 0%, white)"
PASS testGradient("background-image: linear-gradient(to top, black 0%, white)") is "linear-gradient(to top, black 0%, white)"
PASS testGradient("background-image: linear-gradient(10deg, black 0%, white)") is "linear-gradient(10deg, black 0%, white)"
PASS testGradient("background-image: linear-gradient(10deg, black 0%, 25%, white)") is "linear-gradient(10deg, black 0%, 25%, white)"
PASS testGradient("background-image: linear-gradient(10deg, black 0%, 0%, white)") is "linear-gradient(10deg, black 0%, 0%, white)"
PASS testGradient("background-image: linear-gradient(10deg, black 0%, 100%, white)") is "linear-gradient(10deg, black 0%, 100%, white)"
PASS testGradient("background-image: linear-gradient(10deg, black 0%, 50%, green 50%, 50%, white)") is "linear-gradient(10deg, black 0%, 50%, green 50%, 50%, white)"
PASS testGradient("background-image: linear-gradient(black, 10% 20%, white)") is "none"
PASS testGradient("background-image: linear-gradient(black 10% 20% 30%, white)") is "none"
PASS testGradient("background-image: linear-gradient(black 10% 20%, white 100px 200px 300px)") is "none"
PASS testGradient("background-image: linear-gradient(black 10% 20%, white)") is "linear-gradient(black 10%, black 20%, white)"
PASS testGradient("background-image: linear-gradient(black, white 100px 200px)") is "linear-gradient(black, white 100px, white 200px)"
PASS testGradient("background-image: linear-gradient(black 10% 20%, white 100px 200px)") is "linear-gradient(black 10%, black 20%, white 100px, white 200px)"
PASS testGradient("background-image: linear-gradient(black 10% 20%, 50%, white 100px 200px)") is "linear-gradient(black 10%, black 20%, 50%, white 100px, white 200px)"
repeating-linear-gradient

PASS testGradient("background-image: repeating-linear-gradient(black, white)") is "repeating-linear-gradient(black, white)"
PASS testGradient("background-image: repeating-linear-gradient(black, 35%, white)") is "repeating-linear-gradient(black, 35%, white)"
radial-gradient

PASS testGradient("background-image: radial-gradient(white, black)") is "radial-gradient(white, black)"
PASS testGradient("background-image: radial-gradient(at bottom right, white, black)") is "radial-gradient(at right bottom, white, black)"
PASS testGradient("background-image: radial-gradient(farthest-corner, white, black)") is "radial-gradient(white, black)"
PASS testGradient("background-image: radial-gradient(farthest-corner, white, 35%, black)") is "radial-gradient(white, 35%, black)"
PASS testGradient("background-image: radial-gradient(closest-side, white, black)") is "radial-gradient(closest-side, white, black)"
PASS testGradient("background-image: radial-gradient(ellipse, white, black)") is "radial-gradient(white, black)"
PASS testGradient("background-image: radial-gradient(ellipse farthest-corner, white, black)") is "radial-gradient(white, black)"
PASS testGradient("background-image: radial-gradient(ellipse closest-side, white, black)") is "radial-gradient(closest-side, white, black)"
PASS testGradient("background-image: radial-gradient(circle, white, black)") is "radial-gradient(circle, white, black)"
PASS testGradient("background-image: radial-gradient(circle farthest-corner, white, black)") is "radial-gradient(circle, white, black)"
PASS testGradient("background-image: radial-gradient(circle closest-side, white, black)") is "radial-gradient(circle closest-side, white, black)"
PASS testGradient("background-image: radial-gradient(circle closest-side at top, white, black)") is "radial-gradient(circle closest-side at center top, white, black)"
PASS testGradient("background-image: radial-gradient(circle closest-side at top left, white, black)") is "radial-gradient(circle closest-side at left top, white, black)"
PASS testGradient("background-image: radial-gradient(circle closest-side at 10px 20%, white, black)") is "radial-gradient(circle closest-side at 10px 20%, white, black)"
PASS testGradient("background-image: radial-gradient(at 10px 20% circle closest-side, white, black)") is "none"
PASS testGradient("background-image: radial-gradient(circle at 10px 20% circle, white, black)") is "none"
PASS testGradient("background-image: radial-gradient(circle 10px closest-side, white, black)") is "none"
PASS testGradient("background-image: radial-gradient(circle 10%, white, black)") is "none"
PASS testGradient("background-image: radial-gradient(circle 10px 10px, white, black)") is "none"
PASS testGradient("background-image: radial-gradient(ellipse 10%, white, black)") is "none"
PASS testGradient("background-image: radial-gradient(ellipse 10px, white, black)") is "none"
PASS testGradient("background-image: radial-gradient(25%, black)") is "none"
PASS testGradient("background-image: radial-gradient(white, 25%)") is "none"
PASS testGradient("background-image: radial-gradient(white, black, 25%)") is "none"
PASS testGradient("background-image: radial-gradient(25%, white, black)") is "none"
PASS testGradient("background-image: radial-gradient(white,, black)") is "none"
PASS testGradient("background-image: radial-gradient(ellipse 10px, white, 25%, 75%, black)") is "none"
PASS testGradient("background-image: radial-gradient(ellipse 10px 20px, white, black)") is "radial-gradient(10px 20px, white, black)"
PASS testGradient("background-image: radial-gradient(circle 10px, white, black)") is "radial-gradient(10px, white, black)"
PASS testGradient("background-image: radial-gradient(10px 10px at center, white, black)") is "radial-gradient(10px 10px at center center, white, black)"
PASS testGradient("background-image: radial-gradient(10px 10px at center, white, 50%, black)") is "radial-gradient(10px 10px at center center, white, 50%, black)"
PASS testGradient("background-image: radial-gradient(10px 10px at center, white, 0%, black)") is "radial-gradient(10px 10px at center center, white, 0%, black)"
PASS testGradient("background-image: radial-gradient(10px 10px at center, white, 100%, black)") is "radial-gradient(10px 10px at center center, white, 100%, black)"
PASS testGradient("background-image: radial-gradient(black, 10% 20%, white)") is "none"
PASS testGradient("background-image: radial-gradient(black 10% 20% 30%, white)") is "none"
PASS testGradient("background-image: radial-gradient(black 10% 20%, white 100px 200px 300px)") is "none"
PASS testGradient("background-image: radial-gradient(black 10% 20%, white)") is "radial-gradient(black 10%, black 20%, white)"
PASS testGradient("background-image: radial-gradient(black, white 100px 200px)") is "radial-gradient(black, white 100px, white 200px)"
PASS testGradient("background-image: radial-gradient(black 10% 20%, white 100px 200px)") is "radial-gradient(black 10%, black 20%, white 100px, white 200px)"
PASS testGradient("background-image: radial-gradient(black 10% 20%, 50%, white 100px 200px)") is "radial-gradient(black 10%, black 20%, 50%, white 100px, white 200px)"
repeating-radial-gradient

PASS testGradient("background-image: repeating-radial-gradient(white, black)") is "repeating-radial-gradient(white, black)"
PASS testGradient("background-image: repeating-radial-gradient(white, 81%, black)") is "repeating-radial-gradient(white, 81%, black)"
PASS successfullyParsed is true

TEST COMPLETE
